fix: insert new trainer

+++ src/main/kotlin/br/com/ir/pokemon/model/Request.kt
@@ -6,7 +6,7 @@ import java.util.Date
 
 data class TrainerRequest(
 
-    val id: Long? = null,
+    val id: Int? = null,
     val name:String?,
     val birthdate: Date?,
     val created_at:Date?,
@@ -19,13 +19,13 @@ data class TrainerRequest(
     }
 
     data class Builder(
-        var id: Long? = null,
+        var id: Int? = null,
         var name:String? = null,
         var birthdate: Date? = null,
         var created_at:Date? = null,
         var update_at:Date? = null,
         var nickName: String? = null ) {
-            fun id(id:Long) = apply { this.id = id }
+            fun id(id:Int) = apply { this.id = id }
             fun name(name: String) = apply { this.name = name }
             fun birthdate(birthdate: Date) = apply { this.birthdate = birthdate }
             fun created_at(created_at: Date) = apply { this.created_at = created_at }
